requirements.txt
setuptools
wheel
googleapis-common-protos==1.53.0
grpcio==1.38.1
protobuf==3.17.3
jq==1.2.1
python-telegram-bot
simple-chalk==0.1.0
